Comments, Autonomous, and Thread Status Fix
Commented all the code except for constants.cpp.
Autonomous now waits for the box to be lifted and then moves back into
the auto zone. The commented out section of that is supposed to drive to
the right to line up with the next box.
All green pattern added to Arduino code so the lights turn red when
disabled and green when enabled.
Thread status variables did not behave as they were supposed to because
they were dereferenced and saved to local variables instead of being
changed globally, so that is fixed.