Sticks Back to Normal, Lift Fixed, Lift Lights
Buttons are not responding on the aircraft joysticks so the original
joysticks were put back. Field oriented toggle is working and should
make driving easier.

Lift code was changed because it was written using all the grabbing
components.

Lights now scroll blue toward the middle when lifting with thread so it
looks like they are moving up on the robot. Will make them scroll down/
out for when the thread is lowered.